requirements.txt
fastapi[standart]
uvicorn[standart]
PyLTI1p3
flask
pyjwt
dotenv